Fix compilation on 32-bit systems like armv7l
On some 32-bits targets with glibc, time_t still defaults to 32-bit (on
armv7l: a `long`). Meanwhile, a `SQLiteWriter::var_t` can hold either an
`int`, a `long long`, or a `long long unsigned`. Resolve this by casting
to `int64_t`, which is a `long long`.

If compiled for targets where time_t is already a 64-bit integer, this
has no effect.