Add SMTPS and SMTP AUTH support for EmailNotifier, through libcurl
This is intended as a compatible replacement of the existing email
notifier logic: setting 'server' to '10.0.0.2' will still open a SMTP
connection to port 25 on that host, and send the same message as before.
However...

- A hostname can now be given instead of an IP address.
- A port number other than 25 can now be provided.
- If the server hostname is prefixed with 'smtps://', we do TLS.
- If additionally a username and password are given, we do AUTH.

Combined, this looks something like:

    smtps://myuser:mypassword@smtp.eu-west1.example.com:465

The new implementation (like the existing one) does not set a time-out,
but this could be added.